Move `AuthenticationContext` to `hash-graph-sdk`
