+++ apps/hash-api/src/graph/context-types.ts
@@ -1,9 +1,8 @@
 import type { UploadableStorageProvider } from "@local/hash-backend-utils/file-storage";
 import type { TemporalClient } from "@local/hash-backend-utils/temporal";
 import type { GraphApi as GraphApiClient } from "@local/hash-graph-client";
+import type { AuthenticationContext } from "@local/hash-graph-sdk/authentication-context";
 import type { DataSource } from "apollo-datasource";
-
-import type { AuthenticationContext } from "../graphql/authentication-context";
 
 export type GraphApi = GraphApiClient & DataSource;
