+++ apps/hash-api/src/graph/knowledge/primitive/entity/create-entity-hooks.ts
@@ -1,7 +1,7 @@
 import type { VersionedUrl } from "@blockprotocol/type-system";
+import type { AuthenticationContext } from "@local/hash-graph-sdk/authentication-context";
 import type { Entity } from "@local/hash-graph-sdk/entity";
 
-import type { AuthenticationContext } from "../../../../graphql/authentication-context";
 import type { ImpureGraphContext } from "../../../context-types";
 
 export type BeforeCreateEntityHookCallback = (params: {
