+++ apps/hash-api/src/graph/knowledge/primitive/entity/update-entity-hooks.ts
@@ -1,8 +1,8 @@
 import type { VersionedUrl } from "@blockprotocol/type-system";
+import type { AuthenticationContext } from "@local/hash-graph-sdk/authentication-context";
 import type { Entity } from "@local/hash-graph-sdk/entity";
 import type { EntityPropertiesObject } from "@local/hash-graph-types/entity";
 
-import type { AuthenticationContext } from "../../../../graphql/authentication-context";
 import type { ImpureGraphContext } from "../../../context-types";
 
 export type UpdateEntityHookCallback = (params: {
