+++ apps/hash-api/src/graph/knowledge/system-types/file.ts
@@ -4,6 +4,7 @@ import {
   formatFileUrl,
   getEntityTypeIdForMimeType,
 } from "@local/hash-backend-utils/file-storage";
+import type { AuthenticationContext } from "@local/hash-graph-sdk/authentication-context";
 import type { Entity } from "@local/hash-graph-sdk/entity";
 import { generateUuid } from "@local/hash-isomorphic-utils/generate-uuid";
 import { createDefaultAuthorizationRelationships } from "@local/hash-isomorphic-utils/graph-queries";
@@ -19,7 +20,6 @@ import type {
   MutationCreateFileFromUrlArgs,
   MutationRequestFileUploadArgs,
 } from "../../../graphql/api-types.gen";
-import type { AuthenticationContext } from "../../../graphql/authentication-context";
 import type {
   ImpureGraphContext,
   ImpureGraphFunction,
