+++ apps/hash-api/src/graphql/context.ts
@@ -3,12 +3,12 @@ import type { Logger } from "@local/hash-backend-utils/logger";
 import type { SearchAdapter } from "@local/hash-backend-utils/search/adapter";
 import type { TemporalClient } from "@local/hash-backend-utils/temporal";
 import type { VaultClient } from "@local/hash-backend-utils/vault";
+import type { AuthenticationContext } from "@local/hash-graph-sdk/authentication-context";
 
 import type { CacheAdapter } from "../cache";
 import type { EmailTransporter } from "../email/transporters";
 import type { GraphApi } from "../graph/context-types";
 import type { User } from "../graph/knowledge/system-types/user";
-import type { AuthenticationContext } from "./authentication-context";
 
 /**
  * Apollo context object with dataSources. For details see:
