+++ apps/hash-api/src/integrations/linear.ts
@@ -2,10 +2,9 @@ import type { Organization, Team } from "@linear/sdk";
 import { LinearClient } from "@linear/sdk";
 import type { TemporalClient } from "@local/hash-backend-utils/temporal";
 import type { SyncWorkspaceWorkflow } from "@local/hash-backend-utils/temporal-integration-workflow-types";
+import type { AuthenticationContext } from "@local/hash-graph-sdk/authentication-context";
 import type { OwnedById } from "@local/hash-graph-types/web";
 import { generateUuid } from "@local/hash-isomorphic-utils/generate-uuid";
-
-import type { AuthenticationContext } from "../graphql/authentication-context";
 
 export const listTeams = async (params: {
   apiKey: string;
