+++ apps/hash-api/src/seed-data/seed-pages.ts
@@ -1,4 +1,5 @@
 import type { Logger } from "@local/hash-backend-utils/logger";
+import type { AuthenticationContext } from "@local/hash-graph-sdk/authentication-context";
 import type { OwnedById } from "@local/hash-graph-types/web";
 
 import type { ImpureGraphContext } from "../graph/context-types";
@@ -7,7 +8,6 @@ import {
   createPage,
   setPageParentPage,
 } from "../graph/knowledge/system-types/page";
-import type { AuthenticationContext } from "../graphql/authentication-context";
 
 export type PageDefinition = {
   title: string;
