+++ libs/@local/hash-graph-sdk/typescript/src/authentication-context.ts
